+++ src/index.ts
@@ -10,7 +10,8 @@ export {
 	unsubscribeFromPush,
 } from "./client";
 export type { SubscribeResult } from "./client";
-export { sendPushBatch, sendPushNotification, WebPushError } from "./send";
+export { rawPayload, sendPushBatch, sendPushNotification, WebPushError } from "./send";
+export type { RawPushPayload } from "./send";
 export type {
 	Logger,
 	PushPayload,